requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# AI Services
openai==1.12.0
google-generativeai==0.3.2
httpx==0.25.2

# File Processing (for Phase 2)
pdfplumber==0.10.3
python-docx==1.1.0
cloudinary==1.36.0